% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIterativeLSI.R
\name{addIterativeLSI}
\alias{addIterativeLSI}
\title{add iterative LSI results to an SCE}
\usage{
addIterativeLSI(
  x,
  useMatrix = "counts",
  iter = 2,
  preTfIdf = NULL,
  binarize = TRUE,
  corCutOff = 0.75,
  excludeChr = c("chrM", "chrX", "chrY"),
  nDimensions = 30,
  outlierQuantile = c(0.02, 0.98),
  varFeatures = 25000,
  totalFeatures = 5e+05,
  depthCol = "nFrags",
  scaleTo = 10000,
  name = "iterativeLSI",
  seed = 1,
  ...
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{useMatrix}{name of assay to use ("counts")}

\item{iter}{number of LSI iterations (2)}

\item{preTfIdf}{name of precomputed TF-IDF assay, or NULL}

\item{binarize}{binarize counts before TF-IDF? (TRUE)}

\item{corCutOff}{exclude depth-correlated dimensions during clustering (0.75)}

\item{excludeChr}{chromosomes to exclude}

\item{nDimensions}{number of LSI dimensions (30)}

\item{outlierQuantile}{quantiles for excluding depth outliers during LSI}

\item{varFeatures}{number of variable features to select (25000)}

\item{totalFeatures}{number of features considered for feature selection (500000)}

\item{depthCol}{colData column holding sequencing depth ("nFrags")}

\item{scaleTo}{scale factor for normalization (10000)}

\item{name}{name of reducedDim to store the LSI embedding ("iterativeLSI")}

\item{seed}{random seed (1)}

\item{...}{additional arguments}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           SingleCellExperiment with reducedDim(x, name)
}\if{html}{\out{</div>}}
}
\description{
add iterative LSI results to an SCE
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          Runs featureSelectionLSI and stores the final LSI
                   embedding in reducedDim(x, name).
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          featureSelectionLSI
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{          ArchR::addIterativeLSI
}\if{html}{\out{</div>}}
}
