% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLSI.R
\name{addLSI}
\alias{addLSI}
\title{add an LSI result to an SCE (NOTE: this assumes prefiltering was done!)}
\usage{
addLSI(
  x,
  metadataSlot = "TFIDF",
  name = "LSI",
  scaleDims = FALSE,
  corCutOff = 1,
  excludeChr = c("chrM", "chrX", "chrY"),
  nDimensions = 30,
  depth = "nFrags",
  seed = 1,
  ...
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{name}{name of reducedDim to store the LSI embedding ("LSI")}

\item{scaleDims}{Z-scale the dimensions? (TRUE)}

\item{corCutOff}{drop dimensions where cor(LSIdim, nFrag) > corCutOff}

\item{excludeChr}{chroms to exclude by default (chrM, chrX, chrY)}

\item{nDimensions}{number of dimensions for the SVD (30)}

\item{depth}{colData column holding depth for correlations ('nFrags')}

\item{useMatrix}{name of assay to use ("TFIDF")}

\item{subsetLSI}{subset to mcols(x)$usedForLSI? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         SingleCellExperiment with reducedDim(x, name)
}\if{html}{\out{</div>}}
}
\description{
add an LSI result to an SCE (NOTE: this assumes prefiltering was done!)
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        This function assumes that ArchR selected the features!
                 Here we do a simple, noniterative one-pass SVD on TFIDF
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        addTfIdf
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{        ArchR::addIterativeLSI
}\if{html}{\out{</div>}}
}
