% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchRRowZScore.R
\name{ArchRRowZscores}
\alias{ArchRRowZscores}
\title{scale factors (e.g. PCA) to Z-scores, copied from ArchR}
\usage{
ArchRRowZscores(m = NULL, min = -2, max = 2, limit = FALSE)
}
\description{
scale factors (e.g. PCA) to Z-scores, copied from ArchR
}
