% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionLSI.R
\name{.firstPass}
\alias{.firstPass}
\title{perform the first pass of iterative LSI}
\usage{
.firstPass(
  x,
  useMatrix = "counts",
  preTfIdf = NULL,
  binarize = TRUE,
  excludeChr = c("chrM", "chrX", "chrY"),
  nDimensions = 30,
  filterQuantile = 0.995,
  outlierQuantile = c(0.02, 0.98),
  varFeatures = 25000,
  totalFeatures = 5e+05,
  depthCol = "nFrags",
  scaleTo = 10000,
  seed = 1
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{useMatrix}{name of assay to use ("counts")}

\item{preTfIdf}{name of precomputed TF-IDF assay, or NULL}

\item{binarize}{binarize counts before TF-IDF? (TRUE)}

\item{excludeChr}{chromosomes to exclude}

\item{nDimensions}{number of LSI dimensions (30)}

\item{filterQuantile}{quantile for filtering highly accessible features (0.995)}

\item{outlierQuantile}{quantiles for excluding depth outliers during LSI}

\item{varFeatures}{number of variable features to select (25000)}

\item{totalFeatures}{number of features considered for feature selection (500000)}

\item{depthCol}{colData column holding sequencing depth ("nFrags")}

\item{scaleTo}{scale factor for TF-IDF normalization (10000)}

\item{seed}{random seed (1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           list containing LSI, selected features, and TF-IDF result
}\if{html}{\out{</div>}}
}
\description{
perform the first pass of iterative LSI
}
\keyword{internal}
