% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionLSI.R
\name{.getTfIdf}
\alias{.getTfIdf}
\title{retrieve a precomputed TF-IDF matrix}
\usage{
.getTfIdf(
  x,
  selectedFeatures,
  preTfIdf = "TfIdf",
  excludeChr = c("chrX", "chrY", "chrM")
)
}
\arguments{
\item{x}{a SingleCellExperiment containing precomputed TF-IDF values}

\item{selectedFeatures}{features to retain}

\item{preTfIdf}{name of assay containing precomputed TF-IDF values ("TfIdf")}

\item{excludeChr}{chromosomes to exclude}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            list containing the TF-IDF matrix, IDF values, and outlier cells
}\if{html}{\out{</div>}}
}
\description{
retrieve a precomputed TF-IDF matrix
}
\keyword{internal}
