% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionLSI.R
\name{featureSelectionLSI}
\alias{featureSelectionLSI}
\title{perform iterative LSI feature selection}
\usage{
featureSelectionLSI(
  x,
  useMatrix = "counts",
  iter = 2,
  preTfIdf = NULL,
  binarize = TRUE,
  corCutOff = 0.75,
  excludeChr = c("chrM", "chrX", "chrY"),
  nDimensions = 30,
  filterQuantile = 0.995,
  outlierQuantile = c(0.02, 0.98),
  varFeatures = 25000,
  totalFeatures = 5e+05,
  depthCol = "nFrags",
  scaleTo = 10000,
  seed = 1
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{useMatrix}{name of assay to use ("counts")}

\item{iter}{number of LSI iterations (2)}

\item{preTfIdf}{name of precomputed TF-IDF assay, or NULL}

\item{binarize}{binarize counts before TF-IDF? (TRUE)}

\item{corCutOff}{exclude depth-correlated dimensions during clustering (0.75)}

\item{excludeChr}{chromosomes to exclude}

\item{nDimensions}{number of LSI dimensions (30)}

\item{filterQuantile}{quantile for filtering highly accessible features (0.995)}

\item{outlierQuantile}{quantiles for excluding depth outliers during LSI}

\item{varFeatures}{number of variable features to select (25000)}

\item{totalFeatures}{number of features considered for feature selection (500000)}

\item{depthCol}{colData column holding sequencing depth ("nFrags")}

\item{scaleTo}{scale factor for normalization (10000)}

\item{seed}{random seed (1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           list containing LSI, selected features, and TF-IDF result
}\if{html}{\out{</div>}}
}
\description{
perform iterative LSI feature selection
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          Performs iterative LSI following the ArchR approach:
                   LSI, clustering, variable feature selection, and repeated LSI.
}\if{html}{\out{</div>}}
}
\seealso{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          ArchR::addIterativeLSI
}\if{html}{\out{</div>}}
}
