% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTopFeatures.R
\name{findTopFeatures}
\alias{findTopFeatures}
\title{Find top features based on top accessible features}
\usage{
findTopFeatures(
  mat,
  filterQuantile = 0.995,
  varFeatures = 25000,
  totalFeatures = 5e+05
)
}
\arguments{
\item{mat}{a sparse matrix (dgCMatrix) of features x cells}

\item{filterQuantile}{numeric value between 0 and 1; the quantile
of features to filter out from the top of the distribution}

\item{varFeatures}{integer; the number of features to return after filtering}

\item{totalFeatures}{integer; the total number of features in the matrix}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{          character vector of feature names
}\if{html}{\out{</div>}}
}
\description{
Find top features based on top accessible features
}
