% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableFeatures.R
\name{findVariableFeaturesByCluster}
\alias{findVariableFeaturesByCluster}
\title{Find variable features by cluster}
\usage{
findVariableFeaturesByCluster(
  mat,
  clusters,
  nFeatures = 25000,
  scaleTo = 10000,
  method = "var"
)
}
\arguments{
\item{mat}{a sparse matrix (dgCMatrix) of features x cells}

\item{clusters}{a vector of cluster assignments for each cell}

\item{nFeatures}{the number of variable features to return}

\item{scaleTo}{the scale factor for log-normalization (default is 10000)}

\item{method}{the method to use for calculating variability ("var" or "vmr")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     a sparse matrix (dgCMatrix) of the top variable features x cells
}\if{html}{\out{</div>}}
}
\description{
Find variable features by cluster
}
