% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArchRLSIFeatures.R
\name{getArchRLSIFeatures}
\alias{getArchRLSIFeatures}
\title{Get ArchR LSI features}
\usage{
getArchRLSIFeatures(x)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}
}
\value{
A character vector containing the names of features used for LSI.
}
\description{
Extracts the feature names used for latent semantic indexing (LSI) from a
SingleCellExperiment created from an ArchR object.
}
\details{
Features are selected based on the logical \code{usedForLSI} column
in \code{rowData(x)}. An error is raised if \code{usedForLSI} is not present in
the row metadata.
}
