% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndGetMat.R
\name{getChrInd}
\alias{getChrInd}
\title{get chromosome indices}
\usage{
getChrInd(x, chr = c("chrM", "chrX", "chrY"))
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{chr}{chroms to find}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         return indices of target chromosomes
}\if{html}{\out{</div>}}
}
\description{
get chromosome indices
}
\keyword{internal}
