% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLSI.R
\name{getOutliersIdx}
\alias{getOutliersIdx}
\title{Get the outlier cells from a SingleCellExperiment object}
\usage{
getOutliersIdx(x, useMatrix = "TfIdf")
}
\arguments{
\item{x}{A SingleCellExperiment object.}

\item{useMatrix}{The name of the assay to use for outlier detection (default is "TfIdf").}
}
\value{
A logical vector indicating which cells are outliers.
}
\description{
Get the outlier cells from a SingleCellExperiment object
}
