% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{IDF}
\alias{IDF}
\title{Compute inverse document frequencies}
\usage{
IDF(mat)
}
\arguments{
\item{mat}{A sparse count matrix.}
}
\value{
A sparse vector containing one IDF value per feature.
}
\description{
Computes the inverse document frequency (IDF) for each feature using the
definition:
}
\details{
\deqn{
\mathrm{IDF} = \frac{N}{n_i}
}

where \eqn{N} is the number of cells and \eqn{n_i} is the number of cells in
which feature \eqn{i} is observed.
}
\keyword{internal}
