% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{logTFIDF}
\alias{logTFIDF}
\title{Compute log-transformed TF-IDF}
\usage{
logTFIDF(tf, idf, scaleTo = 10000)
}
\arguments{
\item{tf}{A term-frequency normalized sparse matrix.}

\item{idf}{A sparse vector of inverse document frequencies.}

\item{scaleTo}{Scaling factor applied prior to the log transformation. Default same as ArchR.}
}
\value{
A sparse log-transformed TF-IDF matrix.
}
\description{
Multiplies a term-frequency matrix by an inverse document frequency vector
and applies the ArchR log transformation
\eqn{\log(\mathrm{TFIDF} \times scaleTo + 1)}.
}
\keyword{internal}
