% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierByQuantile.R
\name{outlierByQuantile}
\alias{outlierByQuantile}
\title{Identify outlier cells by total accessibility}
\usage{
outlierByQuantile(mat, quantiles = c(0.02, 0.98), excludeZeros = TRUE)
}
\arguments{
\item{mat}{A sparse count matrix.}

\item{quantiles}{Numeric vector, length two, lower and upper quantiles}

\item{excludeZeros}{remove colsum = 0}
}
\value{
logical vector of outlier columns.
}
\description{
Identifies cells whose total counts fall outside user-specified quantiles.
}
