% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneCols.R
\name{pruneCols}
\alias{pruneCols}
\title{Identify low accessibility cells}
\usage{
pruneCols(mat, prune = 1)
}
\arguments{
\item{mat}{A sparse count matrix.}

\item{prune}{Minimum number of features observed within a cell.}
}
\value{
A logical vector indicating which columns should be removed.
}
\description{
Determines which cells observe fewer than a specified number of features
}
