% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{pruneRows}
\alias{pruneRows}
\title{Identify low-frequency features}
\usage{
pruneRows(mat, prune = 1)
}
\arguments{
\item{mat}{A sparse count matrix.}

\item{prune}{Minimum number of cells in which a feature must be observed.}
}
\value{
A logical vector indicating which rows should be removed.
}
\description{
Determines which features occur in fewer than a specified number of cells.
}
\keyword{internal}
