% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDepthCorrelatedDims.R
\name{removeDepthCorrelatedDims}
\alias{removeDepthCorrelatedDims}
\title{Remove LSI dimensions that are correlated with sequencing depth}
\usage{
removeDepthCorrelatedDims(projectedMat, depth, corCutOff = 0.75)
}
\arguments{
\item{projectedMat}{The cell-by-dimension matrix obtained from LSI.}

\item{depth}{A numeric vector representing sequencing depth for each cell.}

\item{corCutOff}{A numeric threshold for correlation; dimensions with correlation above this value will be removed.}
}
\value{
A filtered cell-by-dimension matrix with depth-correlated dimensions removed.
}
\description{
Remove LSI dimensions that are correlated with sequencing depth
}
