% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{TF}
\alias{TF}
\title{Compute term-frequency normalization}
\usage{
TF(mat)
}
\arguments{
\item{mat}{A sparse count matrix.}
}
\value{
A sparse matrix containing term-frequency normalized values.
}
\description{
Computes term frequencies by dividing each nonzero entry by the total count
of its corresponding column.
}
\keyword{internal}
