% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archRtoSCE.R
\name{archRtoSCE}
\alias{archRtoSCE}
\title{turn an ArchR project into a SingleCellExperiment}
\usage{
archRtoSCE(proj, ...)
}
\arguments{
\item{proj}{an ArchRproject}

\item{...}{additional arguments for ArchR::getMatrixFromProject()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         a SingleCellExperiment 
}\if{html}{\out{</div>}}
}
\description{
turn an ArchR project into a SingleCellExperiment
}
\details{
Presumes LSI has been done (this will likely be relaxed soon).
If you are working with single-cell histone or transcription factor
data, you will most likely want to use larger (5kb-50kb) tiles (see
Janssens et al, Nature Protocols, 2024 for sciCUT&Tag benchmarks).
Recent (2026) versions of this function simply cram a lot of ArchR
outputs into metadata()$output for simplicity.
}
