% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLSI.R
\name{calcLSI}
\alias{calcLSI}
\title{Core computational engine: Compute LSI on a matrix}
\usage{
calcLSI(mat, nDimensions = 30, outliers = NULL, scaleDims = FALSE)
}
\arguments{
\item{mat}{Feature-by-cell matrix (e.g. TF-IDF).}

\item{nDimensions}{Number of singular values to calculate.}

\item{outliers}{Logical vector indicating outlier cells to ignore during SVD fitting.}

\item{scaleDims}{Logical; whether to Z-score LSI dimensions across cells.}
}
\value{
A cell-by-dimension projected matrix.
}
\description{
Core computational engine: Compute LSI on a matrix
}
