% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndGetMat.R
\name{filterAndGetMat}
\alias{filterAndGetMat}
\title{Filter features and extract an assay matrix}
\usage{
filterAndGetMat(
  x,
  useMatrix = c("TFIDF", "counts", "TileMatrix"),
  excludeChr = c("chrM", "chrX", "chrY"),
  subsetLSI = FALSE,
  features = NULL,
  binarize = FALSE,
  prune = c(1, 1),
  replaceZeros = TRUE
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{useMatrix}{name of assay to use ("TFIDF")}

\item{excludeChr}{chroms to exclude by default (chrM, chrX, chrY)}

\item{subsetLSI}{subset to mcols(x)$usedForLSI? (FALSE)}

\item{binarize}{binarize matrix? (TRUE)}

\item{prune}{Minimum number of features in a cell must have and minumum of cells containing a feature
c(features,cells)}

\item{replaceZeros}{logical indicating whether filtered features should be replaced with 0 instead}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         filtered assay matrix
}\if{html}{\out{</div>}}
}
\description{
Filter features and extract an assay matrix
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Subsets to features marked by `mcols(x)$usedForLSI`
                 (if requested), removes excluded chromosomes, and
                 returns the requested assay.
}\if{html}{\out{</div>}}
}
\keyword{internal}
