% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndGetMat.R
\name{filterAndGetMat}
\alias{filterAndGetMat}
\title{Filter features and extract an assay matrix}
\usage{
filterAndGetMat(
  x,
  useMatrix = c("TFIDF", "counts", "TileMatrix"),
  excludeChr = c("chrM", "chrX", "chrY"),
  subsetLSI = TRUE
)
}
\arguments{
\item{x}{a SingleCellExperiment, usually from archRtoSCE}

\item{useMatrix}{name of assay to use ("TFIDF")}

\item{excludeChr}{chroms to exclude by default (chrM, chrX, chrY)}

\item{subsetLSI}{subset to mcols(x)$usedForLSI? (TRUE)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         filtered assay matrix
}\if{html}{\out{</div>}}
}
\description{
Filter features and extract an assay matrix
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        Subsets to features marked by `mcols(x)$usedForLSI`
                 (if requested), removes excluded chromosomes, and
                 returns the requested assay.
}\if{html}{\out{</div>}}
}
\keyword{internal}
