% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTF.R
\name{getTF}
\alias{getTF}
\title{Compute term-frequency normalization}
\usage{
getTF(mat)
}
\arguments{
\item{mat}{A sparse count matrix.}
}
\value{
A sparse matrix containing term-frequency normalized values.
}
\description{
Computes term frequencies by dividing each nonzero entry by the total count
of its corresponding column.
}
