% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTfIdf.R
\name{outlierByQuantile}
\alias{outlierByQuantile}
\title{Identify outlier cells by total accessibility}
\usage{
outlierByQuantile(mat, quantiles = c(0.02, 0.98))
}
\arguments{
\item{mat}{A sparse count matrix.}

\item{quantiles}{Numeric vector of length two giving lower and upper
quantiles used to define outliers.}
}
\value{
logical vector of outlier columns.
}
\description{
Identifies cells whose total counts fall outside user-specified quantiles.
}
\keyword{internal}
